% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-addin.R
\name{rs_tar_add_fn}
\alias{rs_tar_add_fn}
\title{rs_tar_add_fn}
\usage{
rs_tar_add_fn(context = rstudioapi::getActiveDocumentContext())
}
\arguments{
\item{context}{from the rstudioapi}
}
\value{
nothing.
}
\description{
Generate a function definition from a function call in RStudio and add to R/functions.R
}
