% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_add_fn.R
\name{tar_add_fn.R}
\alias{tar_add_fn.R}
\title{tar_add_fn.R}
\usage{
tar_add_fn.R(text, index)
}
\value{
Nothing. Modifies your workspace.
}
\description{
Add function to functions.R
}
\details{
Convenience function for adding a function template to functions.R
}
