% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tarflow.R
\name{use_tarflow}
\alias{use_tarflow}
\title{use_tarflow}
\usage{
use_tarflow()
}
\value{
Nothing. Modifies your workspace.
}
\description{
Setup a tarflow project
}
\details{
Creates files and directories according to the tarflow template.
}
