% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarflow_project.R
\name{tarflow_project}
\alias{tarflow_project}
\title{tarflow_project}
\usage{
tarflow_project(path, ...)
}
\arguments{
\item{path}{the path for the new project.}

\item{...}{other project parameters.}
}
\value{
Nothing. Modifies your workspace.
}
\description{
Setup a tarflow project
}
\details{
Creates files and directories according to the tarflow template.
}
