% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tarflow.R
\name{tarflow_project}
\alias{tarflow_project}
\title{use_tarflow}
\usage{
tarflow_project()
}
\value{
Nothing. Modifies your workspace.
}
\description{
Setup a tarflow project
}
\details{
Creates files and directories according to the tarflow template.
}
