% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tarflow.R
\name{use_tarflow}
\alias{use_tarflow}
\title{use_tarflow}
\usage{
use_tarflow(path = getwd(), data = TRUE)
}
\arguments{
\item{path}{where to initialize the tarflow project}

\item{data}{whether or not to include a 'data' directory}
}
\value{
Nothing. Modifies your workspace.
}
\description{
Setup a tarflow project
}
\details{
Creates files and directories according to the tarflow template.
}
